VariantsProcessorConfiguration moved to 'processors' subpackage
