+++ dbsnp-importer/src/main/java/uk/ac/ebi/eva/dbsnpimporter/configuration/processors/SamplesToVariantSourceProcessorConfiguration.java
@@ -1,0 +1,7 @@
+package uk.ac.ebi.eva.dbsnpimporter.configuration.processors;
+
+/**
+ * Created by cyenyxe on 06/12/17.
+ */
+public class SamplesToVariantSourceProcessorConfiguration {
+}
